edit subprocess

+++ convert.py
@@ -77,9 +77,8 @@ def main(args: argparse.Namespace) -> None:
                     fst_key += 1
     os.remove("ngramdata.far")
     os.remove("lm.cnt")
-    os.remove("model.lm")
-    os.remove("model.shrunk.lm")
-
+    if args.ngramshrink:
+        os.remove("model.lm")
 
 # write a function to delete all the temporal files. -> delet lm.cnt. and if ngramshrink, delete model.lm
